api/requirements.txt
flask
flask-restful
mysql-connector-python
Flask-CORS